Renamed feature packages
removed numbers
