--- features/dynamicfeature2/build.gradle.kts
@@ -1,73 +1,0 @@
-plugins {
-    id("com.android.dynamic-feature")
-    id("org.jetbrains.kotlin.android")
-}
-android {
-    namespace = "com.ragnorakdev.dynamicfeature2"
-    compileSdk = 34
-
-    defaultConfig {
-        minSdk = 28
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.5.1"
-    }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-}
-
-dependencies {
-    implementation(project(ModulesApp.app))
-    implementation(UILibraries.activityCompose)
-    implementation(KotlinLibraries.coreKtx)
-    implementation(KotlinLibraries.lifecycleRuntime)
-    implementation(platform(UILibraries.composeBom))
-    implementation(UILibraries.composeUi)
-    implementation(UILibraries.composeGraphics)
-    implementation(UILibraries.composeTooling)
-    implementation(UILibraries.composeMaterial3)
-    implementation(Libraries.koinCore)
-    implementation(Libraries.koinAndroid)
-    implementation(Libraries.koinCompose)
-    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.6.2")
-    implementation("androidx.activity:activity-compose:1.7.2")
-    implementation(platform("androidx.compose:compose-bom:2023.08.00"))
-    implementation("androidx.compose.ui:ui")
-    implementation("androidx.compose.ui:ui-graphics")
-    implementation("androidx.compose.ui:ui-tooling-preview")
-    implementation("androidx.compose.material3:material3")
-    testImplementation(TestingLibraries.jUnit)
-    androidTestImplementation(TestingLibraries.jUnitAndroid)
-    androidTestImplementation(TestingLibraries.espressoCore)
-    androidTestImplementation(TestingLibraries.composeJunit)
-    androidTestImplementation(platform("androidx.compose:compose-bom:2023.08.00"))
-    androidTestImplementation("androidx.compose.ui:ui-test-junit4")
-    debugImplementation(TestingLibraries.uiTooling)
-    debugImplementation(TestingLibraries.uiManifest)
-    debugImplementation("androidx.compose.ui:ui-tooling")
-    debugImplementation("androidx.compose.ui:ui-test-manifest")
-}
