--- features/dynamicfeature2/src/main/AndroidManifest.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:dist="http://schemas.android.com/apk/distribution">
-
-    <dist:module
-        dist:instant="false"
-        dist:title="@string/title_dynamicfeature2">
-        <dist:delivery>
-            <dist:install-time />
-        </dist:delivery>
-
-        <dist:fusing dist:include="true" />
-    </dist:module>
-
-    <application>
-        <activity
-            android:name=".MainSecondFeatureActivity"
-            android:exported="false"
-            android:label="@string/title_activity_main_second_feature"
-            android:theme="@style/Theme.UIRouter" />
-    </application>
-
-</manifest>
