--- features/dynamicfeature3/src/main/java/com/ragnorakdev/dynamicfeature3/ThirdFeatureView.kt
@@ -1,20 +1,0 @@
-package com.ragnorakdev.dynamicfeature3
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-
-@Composable
-fun ThirdFeatureView() {
-    Text(
-        text = "Hello!, I'm Feature 3",
-        modifier = Modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview2() {
-        ThirdFeatureView()
-}
