+++ features/dynamicfeature_one/build.gradle.kts
@@ -1,19 +1,13 @@
 plugins {
-    id("com.android.application")
+    id("com.android.dynamic-feature")
     id("org.jetbrains.kotlin.android")
 }
-
 android {
-    namespace = ApplicationId.applicationId
+    namespace = "com.ragnorakdev.dynamicfeature_one"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = ApplicationId.applicationId
         minSdk = 28
-        targetSdk = 34
-        versionCode = Releases.versionCode
-        versionName = Releases.versionName
-
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
             useSupportLibrary = true
@@ -33,9 +27,6 @@ android {
         sourceCompatibility = JavaVersion.VERSION_17
         targetCompatibility = JavaVersion.VERSION_17
     }
-    kotlinOptions {
-        jvmTarget = "17"
-    }
     buildFeatures {
         compose = true
     }
@@ -47,16 +38,11 @@ android {
             excludes += "/META-INF/{AL2.0,LGPL2.1}"
         }
     }
-    dynamicFeatures += setOf(
-        ModulesApp.featureOne,
-        ModulesApp.featureTwo,
-        ModulesApp.featureThree
-    )
 }
 
 dependencies {
+    implementation(project(ModulesApp.app))
     implementation(project(ModulesApp.uiRouter))
-    implementation(project(ModulesApp.dfManager))
 
     implementation(UILibraries.activityCompose)
     implementation(KotlinLibraries.coreKtx)
