+++ features/dynamicfeature_one/src/main/AndroidManifest.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:dist="http://schemas.android.com/apk/distribution">
+
+    <dist:module
+        dist:instant="false"
+        dist:title="@string/title_dynamicfeature_one">
+        <dist:delivery>
+            <dist:install-time />
+        </dist:delivery>
+
+        <dist:fusing dist:include="true" />
+    </dist:module>
+
+    <application>
+        <activity
+            android:name=".MainFirstFeatureActivity"
+            android:exported="false"
+            android:label="@string/title_activity_main_first_feature"
+            android:theme="@style/Theme.UIRouter" />
+    </application>
+
+</manifest>
